+++ app/src/test/java/com/example/cahier/developer/brushgraph/data/BrushGraphRepositoryTest.kt
@@ -3,6 +3,7 @@ package com.example.cahier.developer.brushgraph.data
 import androidx.ink.brush.BrushFamily
 import com.example.cahier.developer.brushdesigner.data.CustomBrushDao
 import com.example.cahier.developer.brushdesigner.data.CustomBrushEntity
+import com.example.cahier.developer.brushdesigner.data.FakeCustomBrushDao
 import com.example.cahier.core.ui.CahierTextureBitmapStore
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -308,30 +309,4 @@ class BrushGraphRepositoryTest {
 
 }
 
-/** A simple fake for CustomBrushDao for testing. */
-class FakeCustomBrushDao : CustomBrushDao {
-    private val brushes = mutableMapOf<String, CustomBrushEntity>()
-    private val autoSaveFlow = MutableStateFlow<CustomBrushEntity?>(null)
 
-    override fun getAllCustomBrushes(autosaveKey: String): Flow<List<CustomBrushEntity>> {
-        return flowOf(brushes.values.filter { it.name != autosaveKey })
-    }
-
-    override fun getAutoSaveBrush(autosaveKey: String): Flow<CustomBrushEntity?> {
-        return autoSaveFlow.asStateFlow()
-    }
-
-    override suspend fun saveCustomBrush(brush: CustomBrushEntity) {
-        brushes[brush.name] = brush
-        if (brush.name == "__autosave__") {
-            autoSaveFlow.value = brush
-        }
-    }
-
-    override suspend fun deleteCustomBrush(name: String) {
-        brushes.remove(name)
-        if (name == "__autosave__") {
-            autoSaveFlow.value = null
-        }
-    }
-}
